Problem: integer overflow when calculating sizes (may happen e.g. with very large images) Solution: change some types from (unsigned) long to size_t
# Conflicts:
#	dlib/dnn/tensor.h

+++ dlib/binary_search_tree/binary_search_tree_kernel_1.h
@@ -168,7 +168,7 @@ namespace dlib
             );
 
             // functions from the enumerable interface
-            inline unsigned long size (
+            inline size_t size (
             ) const;
 
             bool at_start (
@@ -597,7 +597,7 @@ namespace dlib
         typename mem_manager,
         typename compare
         >   
-    unsigned long binary_search_tree_kernel_1<domain,range,mem_manager,compare>::
+    size_t binary_search_tree_kernel_1<domain,range,mem_manager,compare>::
     size (
     ) const
     {
